added unit test for user services

+++ server/projects/kikbak/src/com/kikbak/dao/ReadOnlyUser2FriendDAO.java
@@ -1,5 +1,6 @@
 package com.kikbak.dao;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 import com.kikbak.dto.User2friend;
@@ -8,5 +9,5 @@ public interface ReadOnlyUser2FriendDAO {
 
 	public User2friend findById(Long id);
 	public Collection<Long> listFriendsForUser(Long userId);
-	public Collection<Long> listFriendsToDelete(Long userId, Collection<Long> facebookIds);
+	public Collection<BigInteger> listFriendsToDelete(Long userId, Collection<Long> facebookIds);
 }
